xivo-aastra: use HTTP to interact with the remote directory
When using HTTPS, the phones use SSLv3, and since SSLv3 support has been
removed on XiVO on the 443 port, the directory is not accesible anymore
with the default configuration on the Aastra.

Other phones (Snom, Yealink) use HTTP to interact with the remote
directory. The Aastra phones downloads its configuration file using
HTTP. The phone doesn't do certificate verification. The remote
directory is accessible to anyone with an allowed IP address.

If someone really wants to use HTTPS to interact with the remote
directory, a custom template can always be used.