xivo-yealink-v80: update templates
Remove the "sip_servers" block. This also fix a bug, i.e. in previous
version, if you had a backup registrar/proxy configured on your XiVO,
and then removed it, the phone would still hold the reference to the
backup registrar/proxy.

Set "account.X.nat.udp_update_enable" to 0, so that by default the phone
stops sending an empty UDP packet to the registrar every 30 seconds.

Set "account.X.fallback.redundancy_type" to 1, i.e. to "successive
registration", which is how XiVO HA works.

Set "features.caller_name_type_on_dialing" to 1, so that the phone
doesn't display the SIP username on the phone when dialing (which is
just random gibberish).

Set "sip.notify_reboot_enable" to 0, so to not force the phone to always
reboot when changing a setting.