xivo-aastra: add language path override
The "language N" option for the 4.1.0 firmware doesn't behave the same
as for the 3.3.1 firmware. The 4.1.0 firmware tries to look for the
language files from the "root" of the HTTP server (e.g.
http://example.org/), instead of trying to look from the URL sent by the
DHCP server (i.e. http://example.org/Aastra).

To me, it looks like a bug from Aastra.