add xivo-polycom-5.4.3 (WIP)
The 5.4 firmware is necessary to support the VVX101 and VVX201 phones.

The plugin is working, but there's still more tests to be done on the
phones, and some things seems a bit broken (the remote directory on the
VVX101 and VVX201, the "remote directory" softkey location, the default
configuration of the second line key on the VVX101).