xivo-snom: handle function keys with "empty string" label
xivo-confd creates configs with these kind of labels, but xivo-provd
plugins expect the 'label' key to NOT be defined if the label is
unspecified, i.e. if a default label must be used.

This was a bit problematic on the D745, since when not specifying a
label, the phone was automatically creating a label from the
"fkey_value" value, and for BLF function keys, the generated label was
actually something like "1001|*8", which was quite ugly.