yealink-v83: undo part of W80 changes
Why:
The W80 in DECT manager mode (DM) does not use a different product name
in both the User-Agent and in the $PN variable. Thus, the DECT manager
advertises itself as a W80B even though it is written W80DM all over the
web interface. This behaviour is different than the W90, which does the
wanted behaviour.